training/requirements.txt
torch>=2.0
numpy
rasterio
pyyaml
wandb
matplotlib
tqdm
